revert: rollback hyprland-qtutils to 119bcb9 due to build failures
The latest changes in hyprland-qtutils (5ffdfc1) cause build script failures
in our current environment. To ensure build stability and continuous
development workflow, we temporarily revert to the previous known working
version (119bcb9).

This change is temporary and will be re-assessed once the issues with
the latest commit are properly addressed.

Ref: PROBLEMATIC_COMMIT=5ffdfc1
Ref: RESTORED_COMMIT=119bcb9